// generated/src/app/api-models/pmb-memb-hist.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbMembHist {
  customerType: string;
  customerId: string;
  invoiceNo: number;
  subscriberId: string;
  personNumber: string;
  seqGpbilId: number;
  membRelationCode: string;
  membLastName: string;
  membFirstName: string;
  membDateOfBirth: Date;
  membGender: string;
  membAge: number;
  premiumStep: string;
  cobraFlag: string;
  medicareStatus: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  rateType: string;
  otherStatusFlag: string;
  premOverrideStep: number;
  premOverrideAmt: number;
  premOverrideCode: string;
  monthlyRate: number;
  seqParentId: number;
}